SEQNG-478 Move the offset related classes to its own file
