+++ modules/edu.gemini.seqexec.web/edu.gemini.seqexec.web.client/src/main/scala/edu/gemini/seqexec/web/client/components/sequence/steps/StepSettings.scala
@@ -3,25 +3,20 @@
 
 package edu.gemini.seqexec.web.client.components.sequence.steps
 
-import edu.gemini.seqexec.model.Model.{FPUMode, Guiding, Instrument, OffsetAxis, Step, StepType, StepState, TelescopeOffset}
+import edu.gemini.seqexec.model.Model.{FPUMode, Guiding, Instrument, Step, StepType, StepState}
 import edu.gemini.seqexec.web.client.actions.FlipBreakpointStep
 import edu.gemini.seqexec.model.enumerations
 import edu.gemini.seqexec.web.client.circuit.{SeqexecCircuit, ClientStatus, StepsTableFocus}
 import edu.gemini.seqexec.web.client.components.SeqexecStyles
-import edu.gemini.seqexec.web.client.components.sequence.steps.OffsetFns._
 import edu.gemini.seqexec.web.client.lenses._
 import edu.gemini.seqexec.web.client.semanticui.elements.label.Label
 import edu.gemini.seqexec.web.client.semanticui.elements.icon.Icon
 import edu.gemini.seqexec.web.client.semanticui.elements.icon.Icon._
 import edu.gemini.seqexec.web.client.semanticui.Size
 import edu.gemini.seqexec.web.client.services.HtmlConstants.iconEmpty
-import edu.gemini.web.client.utils._
-import japgolly.scalajs.react.ScalazReact._
 import japgolly.scalajs.react._
 import japgolly.scalajs.react.component.Scala.Unmounted
 import japgolly.scalajs.react.vdom.html_<^._
-import org.scalajs.dom.CanvasRenderingContext2D
-import org.scalajs.dom.html.Canvas
 
 import scalacss.ScalaCssReact._
 import scalacss.StyleA
@@ -30,141 +25,6 @@ import scalaz.syntax.show._
 import scalaz.syntax.std.option._
 import scalaz.std.anyVal._
 import scalaz.std.string._
-
-/**
-  * Component to draw a grid for the offsets using canvas
-  */
-object OffsetGrid {
-  private val Size = 33.0
-  final case class Props(p: TelescopeOffset.P, q: TelescopeOffset.Q)
-  final case class State(canvas: Option[Canvas])
-
-  private val ST = ReactS.Fix[State]
-
-  @SuppressWarnings(Array("org.wartremover.warts.AsInstanceOf"))
-  def render(props: Props, state: State): Callback = state.canvas.fold(Callback.empty) { c => Callback {
-    // The canvas API is very imperative and stateful but we are inside a Callback!
-    val ctx = c.getContext("2d").asInstanceOf[CanvasRenderingContext2D]
-    c.width = Size.toInt
-    c.height = Size.toInt
-    // First quadrant
-    ctx.fillStyle = if (props.p > TelescopeOffset.P.Zero && props.q > TelescopeOffset.Q.Zero) "darkgray" else "white"
-    ctx.fillRect(0, 0, Size / 2, Size / 2)
-    // Second quadrant
-    ctx.fillStyle = if (props.p < TelescopeOffset.P.Zero && props.q > TelescopeOffset.Q.Zero) "darkgray" else "white"
-    ctx.fillRect(Size / 2, 0, Size / 2, Size / 2)
-    // Third quadrant
-    ctx.fillStyle = if (props.p < TelescopeOffset.P.Zero && props.q < TelescopeOffset.Q.Zero) "darkgray" else "white"
-    ctx.fillRect(Size / 2, Size / 2, Size / 2, Size / 2)
-    // Fourth quadrant
-    ctx.fillStyle = if (props.p > TelescopeOffset.P.Zero && props.q < TelescopeOffset.Q.Zero) "darkgray" else "white"
-    ctx.fillRect(0, Size / 2, Size / 2, Size / 2)
-    // Grid
-    ctx.fillStyle = "black"
-    // Outer border
-    ctx.strokeRect(0, 0, Size, Size)
-    // Inner borders
-    ctx.strokeRect(0, 0, Size / 2, Size / 2)
-    ctx.strokeRect(Size / 2, 0, Size / 2, Size / 2)
-    ctx.strokeRect(0, Size / 2, Size / 2, Size / 2)
-    ctx.strokeRect(Size/2, Size / 2, Size / 2, Size / 2)
-  }}
-
-  private val component = ScalaComponent.builder[Props]("OffsetGrid")
-    .initialState(State(None))
-    .render_P ( p =>
-      <.canvas(
-        SeqexecStyles.offsetGrid,
-        ^.width := Size.toInt.px,
-        ^.height := Size.toInt.px
-      )
-    ).componentWillReceiveProps { ctx =>
-    render(ctx.nextProps, ctx.state)
-  }.componentDidMount { ctx =>
-    // Grab a copy of the canvas
-    val state = State(Some(ctx.getDOMNode.domCast[Canvas]))
-    ctx.runState(ST.set(state)) >> render(ctx.props, state)
-  }.build
-
-  def apply(p: Props): Unmounted[Props, State, Unit] = component(p)
-
-}
-
-/**
-  * Component to display the offsets
-  */
-object OffsetsDisplayCell {
-  final case class Props(offsetsDisplay: OffsetsDisplay, step: Step)
-
-  private val component = ScalaComponent.builder[Props]("OffsetsDisplayCell")
-    .stateless
-    .render_P { p =>
-      <.div( // Column step offset
-        p.offsetsDisplay match {
-          case OffsetsDisplay.DisplayOffsets(offsetWidth) =>
-            OffsetBlock(OffsetBlock.Props(p.step, offsetWidth))
-          case _ => EmptyVdom
-        }
-      )
-    }.build
-
-  def apply(i: Props): Unmounted[Props, Unit, Unit] = component(i)
-}
-
-/**
- * Component to display the offset grid and offset values
- */
-object OffsetBlock {
-  final case class Props(s: Step, offsetWidth: Int)
-  private val component = ScalaComponent.builder[Props]("OffsetValues")
-    .stateless
-    .render_P { p =>
-      val offsetP = telescopeOffsetPO.getOption(p.s).getOrElse(TelescopeOffset.P.Zero)
-      val offsetQ = telescopeOffsetQO.getOption(p.s).getOrElse(TelescopeOffset.Q.Zero)
-
-      <.div(
-        SeqexecStyles.centeredCell,
-        <.div(
-          SeqexecStyles.inlineBlock,
-          SeqexecStyles.offsetCellWrapper,
-          ^.textAlign := "left",
-          OffsetGrid(OffsetGrid.Props(offsetP, offsetQ))
-        ),
-        <.div(
-          SeqexecStyles.inlineBlock,
-          ^.textAlign := "right",
-          <.div(
-            <.div(
-              ^.width := pLabelWidth.px,
-              SeqexecStyles.inlineBlock,
-              offsetAxis(OffsetAxis.AxisP)
-            ),
-            <.div(
-              ^.width := p.offsetWidth.px,
-              SeqexecStyles.inlineBlock,
-              offsetValueFormat(offsetP)
-            )
-          ),
-          <.div(
-              SeqexecStyles.inlineBlock,
-            <.div(
-              ^.width := qLabelWidth.px,
-              SeqexecStyles.inlineBlock,
-              offsetAxis(OffsetAxis.AxisQ)
-            ),
-            <.div(
-              ^.width := p.offsetWidth.px,
-              SeqexecStyles.inlineBlock,
-              offsetValueFormat(offsetQ)
-            )
-          )
-        )
-      )
-    }
-    .build
-
-  def apply(p: Props): Unmounted[Props, Unit, Unit] = component(p)
-}
 
 /**
  * Component to display an icon for the state
